# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "RAG Search API",
        "description": "WealthTech search API across clients and documents. v1: client full-text search, document CRUD + embedding.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/clients/{client_id}/documents": {
            "post": {
                "tags": [
                    "documents"
                ],
                "summary": "Create a document for a client (chunks, enriches, embeds, indexes)",
                "operationId": "create_document_v1_clients__client_id__documents_post",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Client Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentOut"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                }
            }
        },
        "/v1/documents": {
            "get": {
                "tags": [
                    "documents"
                ],
                "summary": "List documents (cursor pagination)",
                "operationId": "list_documents_v1_documents_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "Page size",
                            "default": 20,
                            "title": "Limit"
                        },
                        "description": "Page size"
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Opaque cursor from the previous page",
                            "title": "Cursor"
                        },
                        "description": "Opaque cursor from the previous page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentPage"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/documents/{document_id}": {
            "get": {
                "tags": [
                    "documents"
                ],
                "summary": "Get a document by id",
                "operationId": "get_document_v1_documents__document_id__get",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Document Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentOut"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DocumentCreate": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 1,
                        "title": "Title"
                    },
                    "content": {
                        "type": "string",
                        "maxLength": 1000000,
                        "minLength": 1,
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "content"
                ],
                "title": "DocumentCreate"
            },
            "DocumentListItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "client_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Client Id"
                    },
                    "client_name": {
                        "type": "string",
                        "title": "Client Name"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "client_id",
                    "client_name",
                    "title",
                    "content",
                    "created_at"
                ],
                "title": "DocumentListItem"
            },
            "DocumentOut": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "client_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Client Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "client_id",
                    "title",
                    "content",
                    "created_at"
                ],
                "title": "DocumentOut"
            },
            "DocumentPage": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/DocumentListItem"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "next_cursor": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Next Cursor",
                        "description": "Opaque cursor for the next page"
                    },
                    "has_more": {
                        "type": "boolean",
                        "title": "Has More",
                        "default": false
                    },
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "count"
                ],
                "title": "DocumentPage"
            },
...
            "ErrorDetail": {
                "properties": {
                    "code": {
                        "type": "string",
                        "title": "Code",
                        "examples": [
                            "CLIENT_NOT_FOUND"
                        ]
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "examples": [
                            "Client not found"
                        ]
                    },
                    "resource_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Resource Id",
                        "examples": [
                            "uuid-or-null"
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "title": "ErrorDetail"
            },
            "ErrorResponse": {
                "properties": {
                    "detail": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "ErrorResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}